delete old files and consolidate changes
