--- sds_data_manager/lambda_code/SDSCode/pipeline_lambdas/dependencies/imap_idex_dependencies.yaml
@@ -1,78 +1,0 @@
-# <---- IDEX Dependencies ---->
-# For latest configuration guidelines and requirements, refer to the project documentation.
-
-(l1a, all):
-  - upstream_source: idex
-    upstream_data_type: l0
-    upstream_descriptor: raw
-  - upstream_source: leapseconds
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: spacecraft_clock
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-
-(l1b, sci-1week):
-  - upstream_source: idex
-    upstream_data_type: l1a
-    upstream_descriptor: sci-1week
-  - upstream_source: spin
-    upstream_data_type: spin
-    upstream_descriptor: historical
-  - upstream_source: leapseconds
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: spacecraft_clock
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: ephemeris_reconstructed
-    upstream_data_type: spice
-    upstream_descriptor: historical
-  - upstream_source: ephemeris_predicted
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: attitude_history
-    upstream_data_type: spice
-    upstream_descriptor: historical
-  - upstream_source: planetary_ephemeris
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: imap_frames
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-
-(l2a, sci-1week):
-  - upstream_source: idex
-    upstream_data_type: l1b
-    upstream_descriptor: sci-1week
-  - upstream_source: idex
-    upstream_data_type: ancillary
-    upstream_descriptor: l2a-calibration-curve-yield-params
-  - upstream_source: idex
-    upstream_data_type: ancillary
-    upstream_descriptor: l2a-calibration-curve-t-rise
-
-(l2b, all-1mo):
-  - upstream_source: idex
-    upstream_data_type: l2a
-    upstream_descriptor: sci-1week
-    kickoff_job: false
-  - upstream_source: leapseconds
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: spacecraft_clock
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: idex
-    upstream_data_type: l1b
-    upstream_descriptor: evt
-    kickoff_job: false
