--- sds_data_manager/lambda_code/SDSCode/pipeline_lambdas/dependencies/imap_swapi_dependencies.yaml
@@ -1,220 +1,0 @@
-# <---- SWAPI Dependencies ---->
-# For latest configuration guidelines and requirements, refer to the project documentation.
-
-(l1, hk):
-  - upstream_source: swapi
-    upstream_data_type: l0
-    upstream_descriptor: raw
-  - upstream_source: leapseconds
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: spacecraft_clock
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-
-(l1, sci):
-  - upstream_source: swapi
-    upstream_data_type: l0
-    upstream_descriptor: raw
-  - upstream_source: swapi
-    upstream_data_type: l1a
-    upstream_descriptor: hk
-  - upstream_source: leapseconds
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: spacecraft_clock
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-
-(l2, sci):
-  - upstream_source: swapi
-    upstream_data_type: l1
-    upstream_descriptor: sci
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: esa-unit-conversion
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: lut-notes
-
-(l3a, alpha-sw):
-  - upstream_source: swapi
-    upstream_data_type: l2
-    upstream_descriptor: sci
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: clock-angle-and-flow-deflection-lut
-    kickoff_job: false
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: proton-density-temperature-lut
-    kickoff_job: false
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: alpha-density-temperature-lut
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: energy-gf-pui-lut
-    kickoff_job: false
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: instrument-response-lut
-    kickoff_job: false
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: density-of-neutral-helium-lut
-    kickoff_job: false
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: efficiency-lut
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: helium-inflow-vector
-    kickoff_job: false
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: hydrogen-inflow-vector
-    kickoff_job: false
-
-(l3a, proton-sw):
-  - upstream_source: swapi
-    upstream_data_type: l2
-    upstream_descriptor: sci
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: clock-angle-and-flow-deflection-lut
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: proton-density-temperature-lut
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: alpha-density-temperature-lut
-    kickoff_job: false
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: energy-gf-pui-lut
-    kickoff_job: false
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: instrument-response-lut
-    kickoff_job: false
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: density-of-neutral-helium-lut
-    kickoff_job: false
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: efficiency-lut
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: helium-inflow-vector
-    kickoff_job: false
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: hydrogen-inflow-vector
-    kickoff_job: false
-  - upstream_source: attitude_history
-    upstream_data_type: spice
-    upstream_descriptor: historical
-  - upstream_source: leapseconds
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: imap_frames
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: spacecraft_clock
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: science_frames
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: pointing_attitude
-    upstream_data_type: spice
-    upstream_descriptor: historical
-
-(l3a, pui-he):
-  - upstream_source: swapi
-    upstream_data_type: l2
-    upstream_descriptor: sci
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: clock-angle-and-flow-deflection-lut
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: proton-density-temperature-lut
-    kickoff_job: false
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: alpha-density-temperature-lut
-    kickoff_job: false
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: energy-gf-pui-lut
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: instrument-response-lut
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: density-of-neutral-helium-lut
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: efficiency-lut
-    kickoff_job: false
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: helium-inflow-vector
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: hydrogen-inflow-vector
-  - upstream_source: ephemeris_reconstructed
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: ephemeris_predicted
-    upstream_data_type: spice
-    upstream_descriptor: historical
-  - upstream_source: planetary_ephemeris
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: attitude_history
-    upstream_data_type: spice
-    upstream_descriptor: historical
-  - upstream_source: leapseconds
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: spacecraft_clock
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: imap_frames
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: science_frames
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: pointing_attitude
-    upstream_data_type: spice
-    upstream_descriptor: historical
-
-(l3b, combined):
-  - upstream_source: swapi
-    upstream_data_type: l2
-    upstream_descriptor: sci
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: efficiency-lut
-  - upstream_source: swapi
-    upstream_data_type: ancillary
-    upstream_descriptor: energy-gf-sw-lut
