--- sds_data_manager/lambda_code/SDSCode/pipeline_lambdas/dependencies/imap_swe_dependencies.yaml
@@ -1,97 +1,0 @@
-# <---- SWE Dependencies ---->
-# For latest configuration guidelines and requirements, refer to the project documentation.
-
-(l1a, sci):
-  - upstream_source: swe
-    upstream_data_type: l0
-    upstream_descriptor: raw
-  - upstream_source: leapseconds
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: spacecraft_clock
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-
-(l1b, sci):
-  - upstream_source: swe
-    upstream_data_type: l1a
-    upstream_descriptor: sci
-  - upstream_source: swe
-    upstream_data_type: ancillary
-    upstream_descriptor: l1b-in-flight-cal
-  - upstream_source: swe
-    upstream_data_type: ancillary
-    upstream_descriptor: esa-lut
-  - upstream_source: swe
-    upstream_data_type: ancillary
-    upstream_descriptor: eu-conversion
-  - upstream_source: leapseconds
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: spacecraft_clock
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-
-(l2, sci):
-  - upstream_source: swe
-    upstream_data_type: l1b
-    upstream_descriptor: sci
-  - upstream_source: spin
-    upstream_data_type: spin
-    upstream_descriptor: historical
-
-(l3, sci):
-  - upstream_source: swe
-    upstream_data_type: l1b
-    upstream_descriptor: sci
-  - upstream_source: swe
-    upstream_data_type: l2
-    upstream_descriptor: sci
-  - upstream_source: swe
-    upstream_data_type: ancillary
-    upstream_descriptor: config
-  - upstream_source: swapi
-    upstream_data_type: l3a
-    upstream_descriptor: proton-sw
-  - upstream_source: mag
-    upstream_data_type: l1d
-    upstream_descriptor: norm-dsrf
-  - upstream_source: imap_frames
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: science_frames
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: ephemeris_reconstructed
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: ephemeris_predicted
-    upstream_data_type: spice
-    upstream_descriptor: historical
-  - upstream_source: leapseconds
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: spacecraft_clock
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: pointing_attitude
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: planetary_ephemeris
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
-  - upstream_source: planetary_constants
-    upstream_data_type: spice
-    upstream_descriptor: historical
-    kickoff_job: false
